Update p_enemy.cc
Changes to the formula for Baron of Hell's melee damage, according to Kaiser. Odd thing is that the initial formula existed in the 2020 version of Doom 64, uncorrected. 
Even Kaiser knew that and was annoyed that he didn't catch it himself, even in Doom 64 EX, which is what the 2020 commercial version is based off of.